+++ src/layout/LandingPageLayout.tsx
@@ -1,7 +1,7 @@
 import { Component } from 'react';
 import { Outlet } from 'react-router';
-import Footer from '../components/Footer';
-import Header from '../components/Header';
+import Footer from '../components/layout/Footer';
+import Header from '../components/layout/Header';
 
 export default class LandingPageLayout extends Component {
   render() {
